bug fixes
hmm_picker_predictor.py
1. observation probability matrix size and values are corrected in fwd and bwd models (B_pre_fwd_pick and B_fwd_pick, ans  similar bwd ones)
2. n_states of the fwd and bwd models are corrected
3. n_nodes value setting is corrected and and datatype fixed to int
4. predict_time in the prediction loops are corected - for all models. substate, fwd and bwd

hmm_tray_full_predictor.py
1. missing argument added in calling predict_picking_finish_event

hmmodel.py
1. datatype of emission sequence array is fixed to int (e_seq)

+++ rasberry_des/src/rasberry_des/hmmodel.py
@@ -74,7 +74,7 @@ class HMModel(object):
 
         predict_time - look ahead time (i.e. the time we look into the future based on the last observation)
         """
-        e_seq = numpy.array(obs)
+        e_seq = numpy.array(obs, dtype=numpy.int64)
         # assume the observations were made 1 second apart
         t_seq = numpy.array(range(0,len(e_seq)))
